split test files into separate routes
